Real-Time-Detection-of-Traffic-Violation-main/requirement.txt
numpy==1.24.3
pandas==2.0.3
pillow==10.4.0 
scipy==1.10.1
matplotlib==3.7.5
seaborn==0.13.2
torch==2.4.0+cu118  # 確保是 CUDA 11.8 的版本
torchvision==0.19.0+cu118
torchaudio==2.4.0+cu118
ultralytics==8.3.21
opencv-python==4.10.0.84
filelock==3.13.1
requests==2.32.3
tqdm==4.66.5
pyyaml==6.0.2
typing_extensions==4.11.0
websockets==13.1
